Github-bg-remover/requirements.txt
Flask==3.1.0
Pillow==11.1.0
pyTelegramBotAPI==4.26.0
Requests==2.32.3
python-dotenv==1.0.1